+++ src/Enums/PageGridStyleEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Threls\FilamentPageBuilder\Enums;
+
+enum PageGridStyleEnum: string
+{
+    case NORMAL_GRID = 'normal';
+    case STAGGERED_GRID = 'staggered';
+}