+++ src/PageBuilderServiceProvider.php
@@ -6,7 +6,6 @@ use BezhanSalleh\FilamentLanguageSwitch\LanguageSwitch;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 use Threls\FilamentPageBuilder\Http\Middleware\SetApiLocale;
-use Threls\FilamentPageBuilder\Services\PageImageProcessor;
 
 class PageBuilderServiceProvider extends PackageServiceProvider
 {
@@ -29,7 +28,6 @@ class PageBuilderServiceProvider extends PackageServiceProvider
 
         $this->app['router']->aliasMiddleware('api.locale', SetApiLocale::class);
 
-        $this->app->singleton(PageImageProcessor::class);
     }
 
     public function packageBooted(): void